For review

+++ controllers/cards.js
@@ -1,12 +1,13 @@
+/* eslint-disable no-useless-return */
 const Card = require('../models/card');
 
 module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
-  Card.create({ name, link, owener: req.user._id })
+  Card.create({ name, link, owener: req.user })
     .then((card) => res.send(card))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(400).send({ message: 'Переданы некорректные данные' });
+        res.status(400).send({ message: err.message });
       } else {
         res.status(500).send({ message: 'Ошибка сервера' });
       }
@@ -20,11 +21,21 @@ module.exports.getCards = (req, res) => {
 };
 
 module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(() => Error('NotExist'))
-    .then((card) => res.send({ message: card }))
+  Card.findById(req.params.cardId)
+    .orFail()
+    .then((card) => {
+      if (String(card.owener) === String(req.user._id)) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then(() => res.send({ message: card }));
+      } else {
+        res.status(403).send({
+          message: 'Недостаточно прав',
+        });
+        return;
+      }
+    })
     .catch((err) => {
-      if (err.message === 'NotExist') {
+      if (err.name === 'DocumentNotFoundError') {
         res.status(404).send({ message: 'Объект не найден' });
       } else if (err.name === 'CastError') {
         res.status(400).send({ message: 'Переданы некорректные данные' });
